i40e: fix deleting mac filters
jira LE-1907
Rebuild_History Non-Buildable kernel-3.10.0-693.el7
commit-author Alan Brady <alan.brady@intel.com>
commit a6cb91464b5b09c0ff749e4b01048f93ce450275

There exists a bug in which deleting a mac filter does not actually
occur.  The driver reports that the filter has been deleted with no
error.  The problem occurs because the wrong cmd_flag is passed to the
firmware when deleting the filter.  The firmware reports an error back
to the driver but it is expressly ignored.

This fixes the bug by using the correct flag when deleting a filter.
Without this patch, deleted filters remain in firmware and function as
if they had not been deleted.

Change-ID: I5f22b874f3b83f457702f18f0d5602ca21ac40c3
	Signed-off-by: Alan Brady <alan.brady@intel.com>
	Tested-by: Andrew Bowers <andrewx.bowers@intel.com>
	Signed-off-by: Jeff Kirsher <jeffrey.t.kirsher@intel.com>
(cherry picked from commit a6cb91464b5b09c0ff749e4b01048f93ce450275)
	Signed-off-by: Jonathan Maple <jmaple@ciq.com>