i40e: reopen client after reset
jira LE-1907
Rebuild_History Non-Buildable kernel-3.10.0-693.el7
commit-author Mitch Williams <mitch.a.williams@intel.com>
commit 7be96322a57a3612ea7640e6ff56030a7350fd7f

Allow the client interface to reopen existing clients if they were
closed. This allows clients to recover from reset, which is essential
for supporting VF RDMA. In one instance, the driver was not clearing the
open bit when the client was closed. Add the code to clear this bit so
that the state is accurate and the driver will not attempt to reopen
already-open clients. Remove the ref_cnt variable; it was just getting
in the way and was not being used consistently.

Change-ID: Ic71af4553b096963ac0c56a997f887c9a4ed162d
	Signed-off-by: Mitch Williams <mitch.a.williams@intel.com>
	Signed-off-by: Jeff Kirsher <jeffrey.t.kirsher@intel.com>
(cherry picked from commit 7be96322a57a3612ea7640e6ff56030a7350fd7f)
	Signed-off-by: Jonathan Maple <jmaple@ciq.com>