fsnotify: support overlayfs
jira LE-1907
Rebuild_History Non-Buildable kernel-3.10.0-693.el7
commit-author Aihua Zhang <zhangaihua1@huawei.com>
commit f3fbbb079263bd29ae592478de6808db7e708267

When an event occurs direct it to the overlay inode instead of the real
underlying inode.

This will work even if the file was first on the lower layer and then
copied up, while the watch is there.  This is because the watch is on the
overlay inode, which stays the same through the copy-up.

For filesystems other than overlayfs this is a no-op, except for the
performance impact of an extra pointer dereferece.

Verified to work correctly with the inotify/fanotify tests in LTP.

	Signed-off-by: Aihua Zhang <zhangaihua1@huawei.com>
	Signed-off-by: Miklos Szeredi <mszeredi@redhat.com>
	Cc: Jan Kara <jack@suse.cz>
	Cc: Eric Paris <eparis@redhat.com>
(cherry picked from commit f3fbbb079263bd29ae592478de6808db7e708267)
	Signed-off-by: Jonathan Maple <jmaple@ciq.com>