ixgbe/ixgbevf: Don't use lco_csum to compute IPv4 checksum
jira LE-1907
Rebuild_History Non-Buildable kernel-3.10.0-693.el7
commit-author Alexander Duyck <alexander.h.duyck@intel.com>
commit c54cdc316dbd35695cd54dd425327463c72809e4

In the case of IPIP and SIT tunnel frames the outer transport header
offset is actually set to the same offset as the inner transport header.
This results in the lco_csum call not doing any checksum computation over
the inner IPv4/v6 header data.

In order to account for that I am updating the code so that we determine
the location to start the checksum ourselves based on the location of the
IPv4 header and the length.

Fixes: b83e30104bd9 ("ixgbe/ixgbevf: Add support for GSO partial")
	Signed-off-by: Alexander Duyck <alexander.h.duyck@intel.com>
	Signed-off-by: David S. Miller <davem@davemloft.net>
(cherry picked from commit c54cdc316dbd35695cd54dd425327463c72809e4)
	Signed-off-by: Jonathan Maple <jmaple@ciq.com>