amd-xgbe-phy: Sync PCS and PHY modes after reset
jira LE-1907
Rebuild_History Non-Buildable kernel-3.10.0-693.el7
commit-author Lendacky, Thomas <Thomas.Lendacky@amd.com>
commit 0c95a1faaa155bb49fc4acfec61242e80b8a8502

This patch adds support to sync the states of the PCS and the PHY
after a reset is performed.  If the PCS and the PHY are not in the
same state after reset an extra mode change would be performed. This
extra mode change might not be needed if the PCS and the PHY are
synced up after reset.

	Signed-off-by: Tom Lendacky <thomas.lendacky@amd.com>
	Signed-off-by: David S. Miller <davem@davemloft.net>
(cherry picked from commit 0c95a1faaa155bb49fc4acfec61242e80b8a8502)
	Signed-off-by: Jonathan Maple <jmaple@ciq.com>