be2net: Avoid unnecessary firmware updates of multicast list
jira LE-1907
Rebuild_History Non-Buildable kernel-3.10.0-693.el7
commit-author Sriharsha Basavapatna <sriharsha.basavapatna@broadcom.com>
commit 92fbb1df83ec17f62a46b23507ebb3f06ca10cd3

Eachtime the ndo_set_rx_mode() routine is called, the driver programs the
multicast list in the adapter without checking if there are any changes to
the list. This leads to a flood of RX_FILTER cmds when a number of vlan
interfaces are configured over the device, as the ndo_ gets
called for each vlan interface. To avoid this, we now use __dev_mc_sync()
and __dev_uc_sync() API, but only to detect if there is a change in the
mc/uc lists. Now that we use this API, the code has to be-designed to
issue these API calls for each invocation of the be_set_rx_mode() call.

	Signed-off-by: Sriharsha Basavapatna <sriharsha.basavapatna@broadcom.com>
	Signed-off-by: Sathya Perla <sathya.perla@broadcom.com>
	Signed-off-by: David S. Miller <davem@davemloft.net>
(cherry picked from commit 92fbb1df83ec17f62a46b23507ebb3f06ca10cd3)
	Signed-off-by: Jonathan Maple <jmaple@ciq.com>