igb/igbvf: Don't use lco_csum to compute IPv4 checksum
jira LE-1907
Rebuild_History Non-Buildable kernel-3.10.0-693.el7
commit-author Alexander Duyck <alexander.h.duyck@intel.com>
commit 516165a1e2f22e512a976f8dafd76a22310ccfd9

In the case of IPIP and SIT tunnel frames the outer transport header
offset is actually set to the same offset as the inner transport header.
This results in the lco_csum call not doing any checksum computation over
the inner IPv4/v6 header data.

In order to account for that I am updating the code so that we determine
the location to start the checksum ourselves based on the location of the
IPv4 header and the length.

Fixes: e10715d3e961 ("igb/igbvf: Add support for GSO partial")
	Reported-by: Stephen Rothwell <sfr@canb.auug.org.au>
	Signed-off-by: Alexander Duyck <alexander.h.duyck@intel.com>
	Signed-off-by: David S. Miller <davem@davemloft.net>
(cherry picked from commit 516165a1e2f22e512a976f8dafd76a22310ccfd9)
	Signed-off-by: Jonathan Maple <jmaple@ciq.com>