igb/igbvf: Add support for GSO partial
jira LE-1907
Rebuild_History Non-Buildable kernel-3.10.0-693.el7
commit-author Alexander Duyck <aduyck@mirantis.com>
commit e10715d3e9618901c5ef820a92e6a8e6548b43d3

This patch adds support for partial GSO segmentation in the case of
tunnels.  Specifically with this change the driver an perform segmentation
as long as the frame either has IPv6 inner headers, or we are allowed to
mangle the IP IDs on the inner header.  This is needed because we will not
be modifying any fields from the start of the start of the outer transport
header to the start of the inner transport header as we are treating them
like they are just a block of IP options.

	Signed-off-by: Alexander Duyck <aduyck@mirantis.com>
	Tested-by: Aaron Brown <aaron.f.brown@intel.com>
	Signed-off-by: Jeff Kirsher <jeffrey.t.kirsher@intel.com>
(cherry picked from commit e10715d3e9618901c5ef820a92e6a8e6548b43d3)
	Signed-off-by: Jonathan Maple <jmaple@ciq.com>