qed: PF-VF resource negotiation
jira LE-1907
Rebuild_History Non-Buildable kernel-3.10.0-693.el7
commit-author Yuval Mintz <Yuval.Mintz@qlogic.com>
commit 1cf2b1a971d29d4435a5c462b91d4a22325ec07d

One of the goals of the vf's first message to the PF [acquire]
is to learn about the number of resources available to it [macs, vlans,
etc.]. This is done via negotiation - the VF requires a set of resources,
which the PF either approves or disaproves and sends a smaller set of
resources as alternative. In this later case, the VF is then expected to
either abort the probe or re-send the acquire message with less
required resources.

While this infrastructure exists since the initial submision of qed
SRIOV support, it's in fact completely inoperational - PF isn't really
looking into the resources the VF has asked for and is never going to
reply to the VF that it lacks resources.

This patch addresses this flow, fixing it and allowing the PF and VF
to actually agree on a set of resources.

	Signed-off-by: Yuval Mintz <Yuval.Mintz@qlogic.com>
	Signed-off-by: David S. Miller <davem@davemloft.net>
(cherry picked from commit 1cf2b1a971d29d4435a5c462b91d4a22325ec07d)
	Signed-off-by: Jonathan Maple <jmaple@ciq.com>