qed*: Fix Kconfig dependencies with INFINIBAND_QEDR
jira LE-1907
Rebuild_History Non-Buildable kernel-3.10.0-693.el7
commit-author Yuval Mintz <Yuval.Mintz@caviumnetworks.com>
commit 0189efb8f4f830b9ac7a7c56c0c6e260859e950d
Empty-Commit: Cherry-Pick Conflicts during history rebuild.
Will be included in final tarball splat. Ref for failed cherry-pick at:
ciq/ciq_backports/kernel-3.10.0-693.el7/0189efb8.failed

The qedr driver would require a tristate Kconfig option [to allow
it to compile as a module], and toward that end we've added the
INFINIBAND_QEDR option. But as we've made the compilation of the
qed/qede infrastructure required for RoCE dependent on the option
we'd be facing linking difficulties in case that QED=y or QEDE=y,
and INFINIBAND_QEDR=m.

To resolve this, we seperate between the INFINIBAND_QEDR option
and the infrastructure support in qed/qede by introducing a new
QED_RDMA option which would be selected by INFINIBAND_QEDR but would
be a boolean instead of a tristate; Following that, the qed/qede is
fixed based on this new option so that all config combinations would
be supported.

Fixes: cee9fbd8e2e9 ("qede: add qedr framework")
	Reported-by: Arnd Bergmann <arnd@arndb.de>
	Signed-off-by: Yuval Mintz <Yuval.Mintz@caviumnetworks.com>
	Signed-off-by: David S. Miller <davem@davemloft.net>
(cherry picked from commit 0189efb8f4f830b9ac7a7c56c0c6e260859e950d)
	Signed-off-by: Jonathan Maple <jmaple@ciq.com>

# Conflicts:
#	drivers/net/ethernet/qlogic/Kconfig
#	drivers/net/ethernet/qlogic/qede/Makefile
#	include/linux/qed/qede_roce.h