iw_cxgb4: add fast-path for small REG_MR operations
jira LE-1907
Rebuild_History Non-Buildable kernel-3.10.0-693.el7
commit-author Steve Wise <swise@opengridcomputing.com>
commit 49b53a93a64ab0aaec10851b004297a3ac885433

When processing a REG_MR work request, if fw supports the
FW_RI_NSMR_TPTE_WR work request, and if the page list for this
registration is <= 2 pages, and the current state of the mr is INVALID,
then use FW_RI_NSMR_TPTE_WR to pass down a fully populated TPTE for FW
to write.  This avoids FW having to do an async read of the TPTE blocking
the SQ until the read completes.

To know if the current MR state is INVALID or not, iw_cxgb4 must track the
state of each fastreg MR.  The c4iw_mr struct state is updated as REG_MR
and LOCAL_INV WRs are posted and completed, when a reg_mr is destroyed,
and when RECV completions are processed that include a local invalidation.

This optimization increases small IO IOPS for both iSER and NVMF.

	Signed-off-by: Steve Wise <swise@opengridcomputing.com>
	Signed-off-by: Doug Ledford <dledford@redhat.com>
(cherry picked from commit 49b53a93a64ab0aaec10851b004297a3ac885433)
	Signed-off-by: Jonathan Maple <jmaple@ciq.com>