jbd2: fix r_count overflows leading to buffer overflow in journal recovery
jira LE-1907
Rebuild_History Non-Buildable kernel-3.10.0-693.el7
commit-author Darrick J. Wong <darrick.wong@oracle.com>
commit e531d0bceb402e643a4499de40dd3fa39d8d2e43

The journal revoke block recovery code does not check r_count for
sanity, which means that an evil value of r_count could result in
the kernel reading off the end of the revoke table and into whatever
garbage lies beyond.  This could crash the kernel, so fix that.

However, in testing this fix, I discovered that the code to write
out the revoke tables also was not correctly checking to see if the
block was full -- the current offset check is fine so long as the
revoke table space size is a multiple of the record size, but this
is not true when either journal_csum_v[23] are set.

	Signed-off-by: Darrick J. Wong <darrick.wong@oracle.com>
	Signed-off-by: Theodore Ts'o <tytso@mit.edu>
	Reviewed-by: Jan Kara <jack@suse.cz>
	Cc: stable@vger.kernel.org
(cherry picked from commit e531d0bceb402e643a4499de40dd3fa39d8d2e43)
	Signed-off-by: Jonathan Maple <jmaple@ciq.com>